+++ lib/concerns/wallaby/resourcable.rb
@@ -57,5 +57,41 @@ module Wallaby
       @current_model_class ||=
         controller_to_get(__callee__, :model_class) || Map.model_class_map(current_resources_name)
     end
+
+    # Shorthand of params[:id]
+    # @return [String, nil] ID param
+    def resource_id
+      params[:id]
+    end
+
+    # @return [#each] a collection of all the records
+    def collection
+      @collection ||= paginate current_servicer.collection params
+    end
+
+    # @return [Object] either persisted or unpersisted resource instance
+    def resource
+      @resource ||= begin
+        # white-listed params
+        whitelisted = action_name.in?(SAVE_ACTIONS) ? resource_params : {}
+        if resource_id.present?
+          current_servicer.find resource_id, whitelisted
+        else
+          current_servicer.new whitelisted
+        end
+      end
+    end
+
+    protected
+
+    # To paginate the collection but only when either `page` or `per` param is given,
+    # or HTML response is requested
+    # @see Wallaby::ModelServicer#paginate
+    # @param query [#each]
+    # @return [#each]
+    def paginate(query)
+      paginatable = params[:page] || params[:per] || request.format.symbol == :html
+      paginatable ? current_servicer.paginate(query, params) : query
+    end
   end
 end