+++ lib/helpers/wallaby/index_helper.rb
@@ -9,16 +9,6 @@ module Wallaby
       current_paginator_class.new(model_class, collection, params)
     end
 
-    # @deprecated Use {#paginator_of} instead. It will be removed from 5.3.*
-    # @param model_class [Class] model class
-    # @param collection [#to_a] a collection of all the resources
-    # @param params [ActionController::Parameters] parameters
-    # @return [Wallaby::ModelPaginator]
-    def paginate(model_class, collection, params)
-      Utils.deprecate 'deprecation.paginate', caller: caller
-      paginator_of model_class, collection, params
-    end
-
     # Just a label
     # @return [String]
     def all_label