+++ lib/servicers/wallaby/abstract_model_servicer.rb
@@ -26,12 +26,17 @@ module Wallaby
       end
     end
 
+    attr_reader :model_class, :authorizer
+
+    delegate :user, to: :authorizer
+
     # @param model_class [Class] model class
     # @param authorizer [Wallaby::ModelAuthorizer]
     # @param model_decorator [Wallaby::ModelServicer]
-    def initialize(model_class, authorizer, model_decorator)
+    def initialize(model_class, authorizer, model_decorator = nil)
       @model_class = model_class || self.class.model_class
       raise ArgumentError, I18n.t('errors.required', subject: 'model_class') unless @model_class
+      @model_decorator ||= Map.model_decorator_map model_class
       @authorizer = authorizer
       @provider = Map.service_provider_map(@model_class).new(@model_class, model_decorator)
     end