+++ spec/deprecation/deprecations_spec.rb
@@ -5,7 +5,7 @@ describe 'ResourcePaginator' do
     expect do
       class ProductPaingator < Wallaby::ResourcePaginator
       end
-    end.to output(a_string_starting_with("[DEPRECATION] `Wallaby::ResourcePaginator` will be removed from 5.3.*. Please inherit `Wallaby::ModelPaginator` instead.\n")).to_stderr
+    end.to output(a_string_starting_with("[DEPRECATION] `Wallaby::ResourcePaginator` will be removed from 5.3.*. Please inherit from `Wallaby::ModelPaginator` instead.\n")).to_stderr
   end
 end
 
@@ -27,6 +27,7 @@ describe Wallaby::ResourcesController, type: :controller do
         controller.current_model_service
       end.to output(a_string_starting_with("[DEPRECATION] `current_model_service` will be removed from 5.3.*. Please use `current_servicer` instead.\n")).to_stderr
       expect(controller.current_model_service).to be_a Wallaby::ModelServicer
+      expect(controller.current_model_service).to eq controller.current_servicer
     end
   end
 
@@ -37,18 +38,6 @@ describe Wallaby::ResourcesController, type: :controller do
         controller.authorizer
       end.to output(a_string_starting_with("[DEPRECATION] `authorizer` will be removed from 5.3.*. Please use `current_authorizer` instead.\n")).to_stderr
       expect(controller.authorizer).to be_a Wallaby::ModelAuthorizer
-    end
-  end
-end
-
-describe Wallaby::IndexHelper, type: :helper do
-  describe '#paginate' do
-    it 'shows deprecation message' do
-      controller.params[:resources] = 'products'
-      expect do
-        helper.paginate(Product, Product.where(false), {})
-      end.to output(a_string_starting_with("[DEPRECATION] `paginate` will be removed from 5.3.*. Please inherit `paginator_of` instead.\n")).to_stderr
-      expect(helper.paginate(Product, Product.where(false), {})).to be_a Wallaby::ModelPaginator
     end
   end
 end