+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # TODOs
 
+- `frontend` partial
+- fix metadata[:options] for float form partial
+- fix sort_disabled in link_builder
+- permit has_many through
+- allow access to authorizer and user in servicer
+- extract assign_attributes out from servicer find/new
+- allow access to context and user in authorizer
+
 - pagination check of collection class
 - documentation for interface
 - documentation for localization